[DTSSTCI-1293] Move debugger to main
